New shared package arch
