+++ apps/authentication-service/src/shared/enums.ts
@@ -25,15 +25,3 @@ export enum ConsoleColors {
   BgWhite = "\x1b[47m",
   BgGray = "\x1b[100m",
 }
-
-export enum DefaultPermissions {
-  OPENID = "openid",
-  USERINFO = "userinfo",
-}
-
-export enum WidgetHostPermissions {
-  READ_KEYS = "read:keys",
-  CREATE_KEYS = "create:keys",
-  DELETE_KEYS = "delete:keys",
-  ROTATE_KEYS = "rotate:keys",
-}
