+++ apps/ucp-ui/package.json
@@ -20,6 +20,7 @@
     "@fontsource/work-sans": "^5.0.18",
     "@mui/icons-material": "^5.16.6",
     "@mui/material": "^5.16.6",
+    "@repo/shared-utils": "^0.0.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.23.1"
