+++ apps/ucp-ui/src/App.tsx
@@ -8,13 +8,10 @@ import "@fontsource/work-sans/500.css";
 import "@fontsource/work-sans/700.css";
 import { Experimental_CssVarsProvider as CssVarsProvider } from "@mui/material";
 import { muiTheme } from "./theme";
-import {
-  DefaultPermissions,
-  WidgetHostPermissions,
-} from "./shared/constants/roles";
+import { DefaultPermissions, WidgetHostPermissions } from "@repo/shared-utils";
 
 const App: React.FC = () => {
-  const scope = `${Object.values(DefaultPermissions).join(" ")} ${Object.values(WidgetHostPermissions).join(" ")}`;
+  const scope = `${Object.values(DefaultPermissions as object).join(" ")} ${Object.values(WidgetHostPermissions as object).join(" ")}`;
 
   return (
     <CssVarsProvider theme={muiTheme}>
