+++ apps/ucp-ui/src/shared/constants/roles.ts
@@ -1,15 +1,3 @@
 export enum UserRoles {
   WidgetHost = "WidgetHost",
 }
-
-export enum DefaultPermissions {
-  OPENID = "openid",
-  USERINFO = "userinfo",
-}
-
-export enum WidgetHostPermissions {
-  READ_KEYS = "read:keys",
-  CREATE_KEYS = "create:keys",
-  DELETE_KEYS = "delete:keys",
-  ROTATE_KEYS = "rotate:keys",
-}
