+++ package.json
@@ -27,6 +27,7 @@
   },
   "packageManager": "npm@10.5.0",
   "workspaces": [
-    "apps/*"
+    "apps/*",
+    "packages/*"
   ]
 }
