+++ packages/shared-utils/src/enums.ts
@@ -1,0 +1,11 @@
+export enum DefaultPermissions {
+  OPENID = "openid",
+  USERINFO = "userinfo",
+}
+
+export enum WidgetHostPermissions {
+  READ_KEYS = "read:keys",
+  CREATE_KEYS = "create:keys",
+  DELETE_KEYS = "delete:keys",
+  ROTATE_KEYS = "rotate:keys",
+}
